feat(review): 리뷰 통계 로직 구현 - Template에 reviewCount/ratingSum/avgRating 추가 - BigDecimal 기반 평균 재계산(recalcAvg), HALF_UP 적용 - 스케일 변환 헬퍼(toScaled)로 ×10 정수 누적 일원화 - 리뷰 CUD 시 템플릿 통계 증분 업데이트 연동 - DTO에 reviewCount/avgRating 반환

+++ backend/src/main/java/com/ll/readycode/api/reviews/dto/request/ReviewCreateRequest.java
@@ -4,6 +4,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.DecimalMax;
 import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.NotBlank;
+import java.math.BigDecimal;
 
 public record ReviewCreateRequest(
     @Schema(description = "리뷰 내용", example = "정말 유용한 템플릿이었습니다!", maxLength = 1000)
@@ -12,4 +13,4 @@ public record ReviewCreateRequest(
     @Schema(description = "별점 (1.0 ~ 5.0)", example = "4.5", minimum = "1.0", maximum = "5.0")
         @DecimalMin(value = "1.0", inclusive = true, message = "별점은 1.0 이상이어야 합니다.")
         @DecimalMax(value = "5.0", inclusive = true, message = "별점은 5.0 이하여야 합니다.")
-        Double rating) {}
+        BigDecimal rating) {}
